breeze:registration and login

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,9 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $Notes = Note::orderBy("created_at","desc")->get();
+        $Notes = Note::query()
+        ->where("user_id", request()->user()->id)
+        ->orderBy("created_at","desc")->get();
         // dd($Note);
         return view('note.index',['notes' => $Notes]);
     }
@@ -35,7 +37,7 @@ class NoteController extends Controller
             'note'=>['required','string']
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
         
 
@@ -47,6 +49,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if($note->user_id !== request()->user()->id){
+            abort(403);
+        };
         return view('note.show',['note'=> $note]);
 
     }
@@ -57,6 +62,9 @@ class NoteController extends Controller
     public function edit(Note $note)
     {
         //
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        };
         return view('note.edit',['note'=> $note]);
 
     }
@@ -67,6 +75,10 @@ class NoteController extends Controller
     public function update(Request $request, Note $note)
     {
         //
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
+        ;
         $data = $request->validate([
             'note' =>['required','string']
         ]);
@@ -84,6 +96,10 @@ class NoteController extends Controller
     public function destroy(Note $note)
     {
         //
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
+        ;
         $note->delete();
         return to_route('note.index')->with('message','Note was deleted');
